fix: 修复历史搜索传入参数类型

+++ genshin-map-data/genshin-map-data-core/genshin-map-data-core-model/src/main/java/site/yuanshen/data/vo/HistorySearchVo.java
@@ -36,5 +36,5 @@ public class HistorySearchVo {
      * 记录类型,不传时默认查询全部数据
      */
     @Schema(title = "类型ID(配合记录类型使用)")
-    private List<Integer> id;
+    private List<Long> id;
 }
